Extract Room, User, Others
