+++ packages/liveblocks-core/src/OthersRef.ts
@@ -2,9 +2,10 @@ import { ImmutableRef, merge } from "./ImmutableRef";
 import { asArrayWithLegacyMethods } from "./LegacyArray";
 import { freeze } from "./lib/freeze";
 import { compact, compactObject } from "./lib/utils";
-import type { Others, User } from "./types";
 import type { BaseUserMeta } from "./types/BaseUserMeta";
 import type { JsonObject } from "./types/Json";
+import type { Others } from "./types/Others";
+import type { User } from "./types/User";
 
 type Connection<TUserMeta extends BaseUserMeta> = {
   readonly connectionId: number;
