+++ packages/liveblocks-core/src/types/User.ts
@@ -1,0 +1,33 @@
+import type { BaseUserMeta } from "./BaseUserMeta";
+import type { JsonObject } from "./Json";
+
+/**
+ * Represents a user connected in a room. Treated as immutable.
+ */
+export type User<
+  TPresence extends JsonObject,
+  TUserMeta extends BaseUserMeta
+> = {
+  /**
+   * The connection id of the user. It is unique and increment at every new connection.
+   */
+  readonly connectionId: number;
+  /**
+   * The id of the user that has been set in the authentication endpoint.
+   * Useful to get additional information about the connected user.
+   */
+  readonly id: TUserMeta["id"];
+  /**
+   * Additional user information that has been set in the authentication endpoint.
+   */
+  readonly info: TUserMeta["info"];
+  /**
+   * The user presence.
+   */
+  readonly presence: TPresence;
+
+  /**
+   * False if the user can modify the room storage, true otherwise.
+   */
+  readonly isReadOnly: boolean;
+};
